Added Github Repository Interface
Adding interface for the Github repository
Cleaned up code
Added notification for Rate Limits with Github API

+++ poshsecframework/Web/GithubClient.cs
@@ -11,32 +11,41 @@ using poshsecframework.Strings;
 
 namespace poshsecframework.Web
 {
+    /// <summary>
+    /// A WebRequest wrapper class for the Github API.
+    /// </summary>
     class GithubClient
     {
         #region Private Variables
         HttpWebRequest ghc = null;
         List<String> errors = new List<string>();
+        int ratelimitremaining = 0;
         #endregion
 
         #region Public Methods
-        public Collection<String> GetBranches(String OwnerName, String RepositoryName)
+        /// <summary>
+        /// Requests the JSON response from the Github API for a list of available branches for the given repository.
+        /// </summary>
+        /// <param name="OwnerName">The owner of the repository.</param>
+        /// <param name="RepositoryName">The name of the repository.</param>
+        /// <returns></returns>
+        public Collection<GithubJsonItem> GetBranches(String OwnerName, String RepositoryName)
         {
-            Collection<String> rtn = new Collection<String>();
             Collection<GithubJsonItem> ghi = Get(Path.Combine(StringValue.GithubURI, String.Format(StringValue.BranchFormat, OwnerName, RepositoryName)));
-            if (ghi != null)
-            {
-                foreach (GithubJsonItem gitem in ghi)
-                {
-                    rtn.Add(gitem.Name);
-                }
-            }
-            return rtn;
+            return ghi;
         }
 
-        public void GetArchive(String OwnerName, String RepositoryName, String Branch, String ModuleDirectory)
+        /// <summary>
+        /// Downloads the zipball of the specified repository and branch and unzips it to the Module Directory.
+        /// </summary>
+        /// <param name="OwnerName">The owner of the repository.</param>
+        /// <param name="RepositoryName">The name of the repository.</param>
+        /// <param name="BranchItem">The selected branch item.</param>
+        /// <param name="ModuleDirectory">The target directory for the zipball to be extracted into.</param>
+        public void GetArchive(String OwnerName, String RepositoryName, GithubJsonItem BranchItem, String ModuleDirectory)
         {
             String tmpfile = Path.GetTempFileName();
-            FileInfo savedfile = Download(Path.Combine(StringValue.GithubURI, String.Format(StringValue.ArchiveFormat, OwnerName, RepositoryName, Branch)), tmpfile);
+            FileInfo savedfile = Download(Path.Combine(StringValue.GithubURI, String.Format(StringValue.ArchiveFormat, OwnerName, RepositoryName, BranchItem.Name)), tmpfile);
             if (savedfile != null)
             {
                 try
@@ -47,28 +56,35 @@ namespace poshsecframework.Web
                     {
                         using (za)
                         {
-                            String parentfolder = za.Entries[0].FullName;
-                            System.IO.Compression.ZipFile.ExtractToDirectory(savedfile.FullName, ModuleDirectory);
-                            String newfolder = Path.Combine(ModuleDirectory, parentfolder);
-                            if (Directory.Exists(newfolder))
+                            if (IsValidPSModule(za))
                             {
-                                DirectoryInfo di = new DirectoryInfo(newfolder);
-                                if (Directory.Exists(target))
+                                String parentfolder = za.Entries[0].FullName;
+                                System.IO.Compression.ZipFile.ExtractToDirectory(savedfile.FullName, ModuleDirectory);
+                                String newfolder = Path.Combine(ModuleDirectory, parentfolder);
+                                if (Directory.Exists(newfolder))
                                 {
-                                    Directory.Delete(target, true);
+                                    DirectoryInfo di = new DirectoryInfo(newfolder);
+                                    if (Directory.Exists(target))
+                                    {
+                                        Directory.Delete(target, true);
+                                    }
+                                    di.MoveTo(target);
+                                    di = null;
                                 }
-                                di.MoveTo(target);
-                                di = null;
                             }
-                        }                        
+                            else
+                            {
+                                errors.Add(String.Format(StringValue.InvalidPSModule, RepositoryName, BranchItem.Name));
+                            }
+                        }
                     }
-                    za = null; 
+                    za = null;
                 }
                 catch (Exception ex)
                 {
                     errors.Add(ex.Message);
                 }
-            } 
+            }
             try
             {
                 File.Delete(savedfile.FullName);
@@ -76,13 +92,57 @@ namespace poshsecframework.Web
             catch (Exception dex)
             {
                 errors.Add(dex.Message);
-            
+
             }
             savedfile = null;
         }
         #endregion
 
         #region Private Methods
+        /// <summary>
+        /// Checks to see if the downloaded zipball contains a psd1 file in the root folder. This is required for the powershell modules.
+        /// </summary>
+        /// <param name="zaitem">The ZipArchive item of the zipball.</param>
+        /// <returns>True if the zipball contains a psd1 file. Default is false.</returns>
+        private bool IsValidPSModule(System.IO.Compression.ZipArchive zaitem)
+        {
+            bool rtn = false;
+            int idx = 0;
+            if (zaitem != null)
+            {
+                string rootfolder = zaitem.Entries[0].FullName;
+                do
+                {
+                    System.IO.Compression.ZipArchiveEntry zaentry = zaitem.Entries[idx];
+                    string filename = zaentry.FullName.Replace(rootfolder, "");
+                    if (!filename.Contains("/") && filename.Contains(".psd1"))
+                    {
+                        rtn = true;
+                    }                    
+                    idx++;
+                } while (!rtn && idx < zaitem.Entries.Count());
+            }            
+            return rtn;
+        }
+
+
+        private string GetLastCommit(GithubJsonItem BranchItem)
+        {
+            string rtn = "";
+            Collection<GithubJsonItem> brnchinfo = Get(BranchItem.URL);
+            if (brnchinfo != null && brnchinfo.Count() > 0)
+            { 
+                
+            }
+            return rtn;
+        }
+
+        /// <summary>
+        /// Downloads the specific file from the Github API.
+        /// </summary>
+        /// <param name="uri">The url of the file to download.</param>
+        /// <param name="targetfile">The local filename to use.</param>
+        /// <returns></returns>
         private FileInfo Download(String uri, String targetfile)
         {
             FileInfo rtn = null;
@@ -130,6 +190,11 @@ namespace poshsecframework.Web
             return rtn;
         }
 
+        /// <summary>
+        /// Performs a Get WebRequest to the specified URL and returns a collection of GithubJsonItems.
+        /// </summary>
+        /// <param name="uri">The url for the Get Request.</param>
+        /// <returns></returns>
         private Collection<GithubJsonItem> Get(String uri)
         {
             Collection<GithubJsonItem> rtn = new Collection<GithubJsonItem>();
@@ -140,12 +205,18 @@ namespace poshsecframework.Web
             {
                 ghr = ghc.GetResponse();
             }
+            catch (WebException wex)
+            {
+                ratelimitremaining = GetRateLimitRemaining(wex.Response);
+                errors.Add(uri + ":" + wex.Message);
+            }
             catch (Exception e)
             {
                 errors.Add(uri + ":" + e.Message);
             }
             if (ghr != null)
             {
+                ratelimitremaining = GetRateLimitRemaining(ghr);
                 if (ghr.ContentType == StringValue.ContentTypeJSON)
                 {
                     Stream ghrs = ghr.GetResponseStream();
@@ -179,6 +250,39 @@ namespace poshsecframework.Web
             return rtn;
         }
 
+        private int GetRateLimitRemaining(WebResponse ghr)
+        {
+            int rtn = 0;
+            if (ghr.Headers.Keys.Count > 0)
+            {
+                int idx = -1;
+                bool found = false;
+                do
+                {
+                    idx++;
+                    if (ghr.Headers.Keys[idx] == StringValue.RateLimitKey)
+                    {
+                        found = true;
+                    }
+                } while (!found && idx < ghr.Headers.Keys.Count);
+                if (found)
+                {
+                    string[] vals = ghr.Headers.GetValues(idx);
+                    if (vals != null)
+                    {
+                        string val = vals[0];
+                        int.TryParse(val, out rtn);
+                    }                    
+                }
+            }
+            return rtn;
+        }
+
+        /// <summary>
+        /// Decodes the base64 encoded content.
+        /// </summary>
+        /// <param name="encodedstring">The base64 encoded content.</param>
+        /// <returns></returns>
         private byte[] Decode(String encodedstring)
         {
             byte[] rtn = null;
@@ -198,9 +302,17 @@ namespace poshsecframework.Web
         #endregion
 
         #region Public Properties
+        /// <summary>
+        /// List of any errors that may have occured during any request.
+        /// </summary>
         public List<String> Errors
         {
             get { return errors; }
+        }
+
+        public int RateLimitRemaining
+        {
+            get { return ratelimitremaining; }
         }
         #endregion        
     }
